refactor: move check-box.tsx to check-box/check-box.tsx
