+++ src/components/certificates/certificates.tsx
@@ -1,7 +1,7 @@
 import { ChangeEvent, useState } from "react";
 import { Months } from "./months/months";
 import { Years } from "./years/years";
-import { CheckBox } from "./check-box";
+import { CheckBox } from "./check-box/check-box";
 import { useForm } from "react-hook-form";
 import { Details } from "../details/details";
 import { CertificateState } from "./types";
